+++ images/api/src/Schema.js
@@ -1,7 +1,8 @@
 const uuidV1 = require('uuid/v1');
 const moment = require('moment');
 
-const {checkToken} = require('./helpers/auth')
+const { checkToken } = require('./helpers/auth')
+const { getSchema } = require('./helpers/getScheme');
 
 class Schema {
 
@@ -58,7 +59,7 @@ class Schema {
       
         const id = await pg('schema').insert(request).returning('id');
         req.params.uuid = request['uuid'];
-        this.getSchema(pg, req, res);
+        getSchema(pg, req.params.uuid, res);
       }, res)
     })
 
@@ -137,7 +138,7 @@ class Schema {
               }
             }
           }
-          this.getSchema(pg, req, res);
+          getSchema(pg, req.params.uuid, res);
         }, res)
       } 
       else {
@@ -147,7 +148,7 @@ class Schema {
     })
 
     app.get('/schema/:uuid', async (req, res, next) => {
-      this.getSchema(pg, req, res)
+      getSchema(pg, req.params.uuid, res)
     });
 
     app.get('/schema', async (req, res, next) => {
@@ -179,115 +180,6 @@ class Schema {
     })
   }
 
-  async getSchema(pg, req, res) {
-    let result = {};
-
-
-    if (req.params.uuid == "new"){
-      result = {
-        "uuid": "new",
-        "title": "A new scheme",
-        "headers": ['A string', 11, 'Click me!', '13:00', 14, 'Tomorrow'],
-        "rows": [
-          {
-            "name": "New row",
-            "cells": [{"current": [],"max": 0},{"current": [],"max": 0},{"current": [],"max": 0},{"current": [],"max": 0},{"current": [],"max": 0},{"current": [],"max": 0}]
-          },
-          {
-            "name": "Another one, I swear",
-            "cells": [{"current": [],"max": 0},{"current": [],"max": 0},{"current": [],"max": 0},{"current": [],"max": 0},{"current": [],"max": 0},{"current": [],"max": 0}]
-          }
-        ],
-        "status": 0,
-        "publication": {
-          "from": new Date(),
-          "to": new Date()
-        },
-        "roles": [],
-        "consumer": {
-          "index": null
-        }
-      };
-      res.status(200).send(result);
-    }
-    else {
-      const rowstructure = [];
-
-
-      let answers = [];
-      await pg.select("answers.cellID", "answers.participant", "answers.created_at", "answers.activated", "answers.updated_at", "cells.row", "cells.col").table("answers").where({"answers.tableID": req.params.uuid}).join('cells', 'cells.uuid', '=', 'answers.cellID').then(function(a) {
-        answers = a;
-      })
-
-
-      // @TODO: make sure param 'uuid' is of type uuid
-      await pg.select()
-        .table('schema')
-        .where({'schema.uuid': req.params.uuid})
-        .leftJoin('cells', 'schema.id', '=', 'cells.tableID')
-        .then(async function (r) {
-          if (r.length > 0) {
-            result['uuid'] = req.params.uuid;
-            result['title'] = r[0].title;
-            result['consumer'] = r[0].consumer;
-            result['status'] = r[0].published;
-            result['roles'] = r[0].roles.roles.map((role) => { return role.type });
-            result['publication'] = {
-              from: moment(r[0].opens).format(),
-              to: moment(r[0].closes).format()
-            };
-            const temp = [];
-            Object.keys(r[0].headers).map((key, index) => {
-              temp.push(key);
-            });
-            result['headers'] = temp;
-
-
-            result['created_at'] = r[0].created_at;
-            result['updated_at'] = r[0].updated_at;
-
-            Object.keys(r[0].rows).map((key, index) => {
-              const found = [];
-              for (let i = 0; i < r.length; i++) {
-                if (r[i]['row'] === key) {
-                  const num = answers.filter(answer => answer.cellID === r[i].uuid);
-                  found.push({
-                    max: r[i].max,
-                    current: num,
-                    col: r[i].col,
-                    uuid: r[i].uuid
-                  })
-                }
-              }
-
-              const intermediary = [];
-
-              for (let i = 0; i < temp.length; i++) {
-                for (let j = 0; j < found.length; j++) {
-                  if (found[j].col === temp[i]) {
-                    intermediary.push(found[j])
-                  }
-                }
-              }
-
-              rowstructure.push({
-                name: key,
-                cells: intermediary
-              });
-            });
-
-
-            result['rows'] = rowstructure;
-
-            result['answers'] = answers;
-            res.status(200).send(result);
-          }
-          else {
-            res.status(404).send({error: 'something went wrong'});
-          }
-        });
-    }
-  }
 
 }
